requirements.txt
alembic==1.6.5
Authlib==0.15.4
cachelib==0.3.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
cryptography==3.4.8
ecdsa==0.17.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Migrate==3.0.1
Flask-OAuth==0.12
Flask-Script==2.0.6
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
Flask-WTF==0.15.1
greenlet==1.1.0
gunicorn==20.1.0
httplib2==0.19.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jose==1.0.0
Mako==1.1.4
MarkupSafe==2.0.1
oauth2==1.9.0.post1
psycopg2-binary==2.9.1
pyasn1==0.4.8
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
python-jose==3.3.0
requests==2.26.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.20
urllib3==1.26.6
Werkzeug==2.0.1
WTForms==2.3.3
